Refactor: Extract test helper logic
